First working version to implement Github PR #19
