+++ bidscoin/plugins/nibabel2bids.py
@@ -13,12 +13,7 @@ from nibabel.testing import data_path
 from bids_validator import BIDSValidator
 from typing import Union
 from pathlib import Path
-try:
-    from bidscoin import bids
-except ImportError:
-    import sys
-    sys.path.append(str(Path(__file__).parents[1]))                 # This should work if bidscoin was not pip-installed
-    import bids
+from bidscoin import bids
 
 LOGGER = logging.getLogger(__name__)
