+++ bidscoin/plugins/phys2bidscoin.py
@@ -17,12 +17,7 @@ import pandas as pd
 import ast
 from pathlib import Path
 from functools import lru_cache
-try:
-    from bidscoin import bids
-except ImportError:
-    import sys
-    sys.path.append(str(Path(__file__).parents[1]))             # This should work if bidscoin was not pip-installed
-    import bids
+from bidscoin import bids
 
 LOGGER = logging.getLogger(__name__)
