+++ bidscoin/utilities/dicomsort.py
@@ -1,10 +1,5 @@
 #!/usr/bin/env python3
-"""
-Sorts and/or renames DICOM files into local subfolders, e.g. with 3-digit SeriesNumber-SeriesDescription
-folder names (i.e. following the same listing as on the scanner console)
-
-Supports flat DICOM as well as multi-subject/session DICOMDIR file structures.
-"""
+"""Sorts and/or renames DICOM files into local subfolders (See also cli/_dicomsort.py)"""
 
 import re
 import logging
@@ -12,12 +7,11 @@ import uuid
 from pydicom import fileset
 from pathlib import Path
 from typing import List
-try:
-    from bidscoin import bcoin, bids
-except ImportError:
+from importlib.util import find_spec
+if find_spec('bidscoin') is None:
     import sys
-    sys.path.append(str(Path(__file__).parents[1]))             # This should work if bidscoin was not pip-installed
-    import bcoin, bids
+    sys.path.append(str(Path(__file__).parents[2]))
+from bidscoin import bcoin, bids
 
 LOGGER = logging.getLogger(__name__)
 
@@ -215,31 +209,12 @@ def sortsessions(sourcefolder: Path, subprefix: str='', sesprefix: str='', folde
 def main():
     """Console script usage"""
 
-    # Parse the input arguments and run the sortsessions(args)
-    import argparse
-    import textwrap
-
-    class CustomFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter): pass
-
-    parser = argparse.ArgumentParser(formatter_class=CustomFormatter,
-                                     description=textwrap.dedent(__doc__),
-                                     epilog='examples:\n'
-                                            '  dicomsort sub-011/ses-mri01\n'
-                                            '  dicomsort sub-011/ses-mri01/DICOMDIR -n {AcquisitionNumber:05d}_{InstanceNumber:05d}.dcm\n'
-                                            '  dicomsort myproject/raw/DICOMDIR --subprefix pat^ --sesprefix\n ')
-    parser.add_argument('dicomsource',          help='The root folder containing the dicomsource/[sub/][ses/] dicomfiles or the DICOMDIR file')
-    parser.add_argument('-i','--subprefix',     help='Provide a prefix string for recursive sorting of dicomsource/subject subfolders (e.g. "sub-")')
-    parser.add_argument('-j','--sesprefix',     help='Provide a prefix string for recursive sorting of dicomsource/subject/session subfolders (e.g. "ses-")')
-    parser.add_argument('-f','--folderscheme',  help='Naming scheme for the sorted DICOM Series subfolders. Follows the Python string formatting syntax with DICOM field names in curly bracers with an optional number of digits for numeric fields. Sorting in subfolders is skipped when an empty folderscheme is given (but note that renaming the filenames can still be performed)', default='{SeriesNumber:03d}-{SeriesDescription}')
-    parser.add_argument('-n','--namescheme',    help='Optional naming scheme that can be provided to rename the DICOM files. Follows the Python string formatting syntax with DICOM field names in curly bracers with an optional number of digits for numeric fields. Use e.g. "{PatientName}_{SeriesNumber:03d}_{SeriesDescription}_{AcquisitionNumber:05d}_{InstanceNumber:05d}.dcm" or "{InstanceNumber:05d}_{SOPInstanceUID}.IMA" for default names')
-    parser.add_argument('-p','--pattern',       help='The regular expression pattern used in re.match(pattern, dicomfile) to select the dicom files', default=r'.*\.(IMA|dcm)$')
-    parser.add_argument('--force',              help='Sort the DICOM data even the DICOM fields of the folder/name scheme are not in the data', action='store_true')
-    parser.add_argument('-d','--dryrun',        help='Only print the dicomsort commands without actually doing anything', action='store_true')
-    args = parser.parse_args()
+    from bidscoin.cli._dicomsort import get_parser
 
     # Set-up logging
     bcoin.setup_logging()
 
+    args = get_parser().parse_args()
     sortsessions(sourcefolder = args.dicomsource,
                  subprefix    = args.subprefix,
                  sesprefix    = args.sesprefix,
