+++ bidscoin/utilities/physio2tsv.py
@@ -1,17 +1,14 @@
 #!/usr/bin/env python3
-"""
-Reads and plots active (i.e. non-zero) signals from SIEMENS advanced physiological log / DICOM files
+""" Reads and writes SIEMENS advanced physiological log / DICOM files (See also cli/_physio2tsv.py)"""
 
-This function expects to find either a combination of individual logfiles (*_ECG.log, *_RESP.log,
-*_PULS.log, *_EXT.log, *_Info.log) generated by >=R013 sequences, or a single encoded "_PHYSIO" DICOM
-file generated by >=R015 sequences.
-"""
-
-import logging, coloredlogs
-try:
-    import bidscoin.utilities.physio as ph
-except ImportError:
-    import physio as ph                 # This should work if bidscoin was not pip-installed
+import logging
+import coloredlogs
+from importlib.util import find_spec
+if find_spec('bidscoin') is None:
+    import sys
+    from pathlib import Path
+    sys.path.append(str(Path(__file__).parents[2]))
+import bidscoin.utilities.physio as ph
 
 # Set-up logging
 LOGGER = logging.getLogger(__name__)
@@ -22,19 +19,9 @@ if not LOGGER.handlers:
 def main():
     """Console script usage"""
 
-    # Parse the input arguments and run physio2tsv(args)
-    import argparse
+    from bidscoin.cli._physio2tsv import get_parser
 
-    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter,
-                                     description=__doc__,
-                                     epilog='examples:\n'
-                                            '  physio2tsv myproject/sub-001/MR000000.dcm myproject/bids/sub-001/func/sub-001_physio\n'
-                                            '  physio2tsv myproject/sub-001/Physio_20200428_142451_007e910e-02d9-4d7a-8fdb-8e3568be8322 myproject/bids/sub-001/func/sub-001_physio\n\n'
-                                            '@author: Marcel Zwiers\n ')
-    parser.add_argument('physiofile', help="Either the fullpath of the DICOM file or the basename of the PHYSIO logfiles (fullpath without suffix and file extension, e.g. 'foo/bar/Physio_DATE_TIME_UUID'")
-    parser.add_argument('tsvfile',    help="The fullpath of the BIDS filenames, e.g. 'foo/bids/sub-001/func/sub-001_physio'")
-    args = parser.parse_args()
-
+    args   = get_parser().parse_args()
     physio = ph.readphysio(args.physiofile)
     ph.physio2tsv(physio, args.tsvfile)
