+++ bidscoin/utilities/plotphysio.py
@@ -1,17 +1,14 @@
 #!/usr/bin/env python3
-"""
-Reads and plots active (i.e. non-zero) signals from SIEMENS advanced physiological log / DICOM files
+""" Plots SIEMENS advanced physiological log / DICOM files (See also cli/_plotphysio.py)"""
 
-This function expects to find either a combination of individual logfiles (*_ECG.log, *_RESP.log,
-*_PULS.log, *_EXT.log, *_Info.log) generated by >=R013 sequences, or a single encoded "_PHYSIO" DICOM
-file generated by >=R015 sequences.
-"""
-
-import logging, coloredlogs
-try:
-    import bidscoin.utilities.physio as ph
-except ImportError:
-    import physio as ph                 # This should work if bidscoin was not pip-installed
+import logging
+import coloredlogs
+from importlib.util import find_spec
+if find_spec('bidscoin') is None:
+    import sys
+    from pathlib import Path
+    sys.path.append(str(Path(__file__).parents[2]))
+import bidscoin.utilities.physio as ph
 
 # Set-up logging
 LOGGER = logging.getLogger(__name__)
@@ -22,19 +19,9 @@ if not LOGGER.handlers:
 def main():
     """Console script usage"""
 
-    # Parse the input arguments and run plotphysio(args)
-    import argparse
+    from bidscoin.cli._plotphysio import get_parser
 
-    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter,
-                                     description=__doc__,
-                                     epilog='examples:\n'
-                                            '  plotphysio myproject/sub-001/MR000000.dcm\n'
-                                            '  plotphysio -s 2000 myproject/sub-001/Physio_20200428_142451_007e910e-02d9-4d7a-8fdb-8e3568be8322\n\n'
-                                            '@author: Marcel Zwiers\n ')
-    parser.add_argument('filename',           help="Either the fullpath of the DICOM file or the basename of the PHYSIO logfiles (fullpath without suffix and file extension, e.g. 'foo/bar/Physio_DATE_TIME_UUID'")
-    parser.add_argument('-s','--showsamples', help='The nr of plotted samples of the physiological traces (default: 1000, nothing is plotted if 0)', default=1000, type=int)
-    args = parser.parse_args()
-
+    args   = get_parser().parse_args()
     physio = ph.readphysio(args.filename)
     ph.plotphysio(physio, args.showsamples)
