+++ tests/test_bids.py
@@ -3,17 +3,11 @@ import pytest
 import shutil
 import re
 import json
+import ruamel.yaml.comments
 from pathlib import Path
 from nibabel.testing import data_path
-
-import ruamel.yaml.comments
 from pydicom.data import get_testdata_file
-try:
-    from bidscoin import bcoin, bids
-except ImportError:
-    import sys
-    sys.path.append(str(Path(__file__).parents[1]/'bidscoin'))         # This should work if bidscoin was not pip-installed
-    import bcoin, bids
+from bidscoin import bcoin, bids
 
 bcoin.setup_logging()
