+++ tests/test_bidscoiner.py
@@ -1,18 +1,12 @@
 import json
-from pathlib import Path
-try:
-    from bidscoin import bcoin, bidsmapper, bidscoiner
-except ImportError:
-    import sys
-    sys.path.append(str(Path(__file__).parents[1]/'bidscoin'))      # This should work if bidscoin was not pip-installed
-    import bcoin, bidsmapper, bidscoiner
+from bidscoin import bcoin, bidsmapper, bidscoiner, bidsmap_template
 
 bcoin.setup_logging()
 
 
 def test_bidscoiner(raw_dicomdir, bids_dicomdir, bidsmap_dicomdir):
     if not bidsmap_dicomdir.is_file():
-        bidsmapper.bidsmapper(raw_dicomdir, bids_dicomdir, bidsmap_dicomdir, bcoin.bidsmap_template, [], 'Doe^', '*', unzip='', noeditor=True, force=True)
+        bidsmapper.bidsmapper(raw_dicomdir, bids_dicomdir, bidsmap_dicomdir, bidsmap_template, [], 'Doe^', '*', unzip='', noeditor=True, force=True)
         try:
             (bidsmap_dicomdir.parent/'bidsmapper.errors').unlink(missing_ok=True)
         except Exception:
