+++ tests/test_utilities.py
@@ -2,15 +2,8 @@ import shutil
 import csv
 from pydicom.data import get_testdata_file
 from pathlib import Path
-try:
-    from bidscoin import bcoin
-    from bidscoin.utilities import dicomsort, rawmapper, bidsparticipants
-except ImportError:
-    import sys
-    sys.path.append(str(Path(__file__).parents[1]/'bidscoin'))         # This should work if bidscoin was not pip-installed
-    sys.path.append(str(Path(__file__).parents[1]/'bidscoin'/'utilities'))
-    import bcoin
-    import dicomsort, rawmapper, bidsparticipants
+from bidscoin import bcoin
+from bidscoin.utilities import dicomsort, rawmapper, bidsparticipants
 
 bcoin.setup_logging()
