+++ tox.ini
@@ -3,7 +3,9 @@ envlist = py38, py39, py310, py311
 
 [testenv]
 extras = all
-deps = pytest
+deps =
+    pytest
+    argparse-manpage[setuptools]
 commands = pytest
 
 [gh-actions]
